TP3/flask-app/requirements.txt
flask
pandas
xmlschema
lxml
requests
grpcio
grpcio-tools
requests